requirements .txt

streamlit
numpy
Pillow
